refactor(apps/discord-bot): update db functions import path
